Ajout PV vaisseau + réparation + amélioration des ordres.

+++ Script/resertunivers.php
@@ -2,21 +2,32 @@
 session_start();
 include("../include/BDDconnection.php");
 
-$reqnomtable = $bda->query("SELECT distinct TABLE_SCHEMA, TABLE_NAME from INFORMATION_SCHEMA.COLUMNS WHERE TABLE_SCHEMA = 'autre' OR TABLE_SCHEMA = 'gamer'");
-while ($repnomtable  = $reqnomtable ->fetch())
+//echo $_POST['confirmer'] . '</br>' ;
+
+if (!isset($_POST['perdreressources']))
+    {
+    header('Location: ../administration.php?message=42');
+    exit(); 
+    }
+
+else
 	{
-	if ($repnomtable['TABLE_NAME'] != 'afaire')
+	$reqnomtable = $bda->query("SELECT distinct TABLE_SCHEMA, TABLE_NAME from INFORMATION_SCHEMA.COLUMNS WHERE TABLE_SCHEMA = 'autre' OR TABLE_SCHEMA = 'gamer'");
+	while ($repnomtable  = $reqnomtable ->fetch())
 		{
-		if ($repnomtable['TABLE_SCHEMA'] == 'autre')  
+		if ($repnomtable['TABLE_NAME'] != 'afaire')
 			{
-			$reqtruncateautre = $bda->prepare('TRUNCATE TABLE '.$repnomtable['TABLE_NAME'].'');
-			$reqtruncateautre->execute(array());
-			}	
-		elseif ($repnomtable['TABLE_SCHEMA'] == 'gamer')
-			{
-			$reqtruncateautre = $bdg->prepare('TRUNCATE TABLE '.$repnomtable['TABLE_NAME'].'');
-			$reqtruncateautre->execute(array());
-			}	
+			if ($repnomtable['TABLE_SCHEMA'] == 'autre')  
+				{
+				$reqtruncateautre = $bda->prepare('TRUNCATE TABLE '.$repnomtable['TABLE_NAME'].'');
+				$reqtruncateautre->execute(array());
+				}	
+			elseif ($repnomtable['TABLE_SCHEMA'] == 'gamer')
+				{
+				$reqtruncateautre = $bdg->prepare('TRUNCATE TABLE '.$repnomtable['TABLE_NAME'].'');
+				$reqtruncateautre->execute(array());
+				}	
+			}
 		}
 	}
 
@@ -27,7 +38,6 @@ session_destroy();
 // Suppression des cookies de connexion automatique
 setcookie("id", 0, time(), "/"); 
 setcookie("pass", 0, time(), "/");
-
 
 header('Location: ../acceuil.php');
 ?>
